split agent worker out of agent
